+++ CHANGELOG.md
@@ -8,6 +8,11 @@
     Polygon, Polyline, PolylineTo, PolyPolygon, PolyPolyline, PtInRegion,
     RectInRegion, RoundRect, SetStretchBltMode, StretchBlt
   - kernel32: GlobalAlloc, GlobalFree, GlobalLock, GlobalUnlock
+  - user32: InflateRect, IntersectRect, InvalidateRgn, InvertRect, IsRectEmpty,
+    LockWindowUpdate, MapWindowPoints, MonitorFromRect, OffsetRect,
+    PaintDesktop, PtInRect, RedrawWindow, ScreenToClient, SetRect, SetRectEmpty,
+    SetWindowRgn, SubtractRect, TabbedTextOut, UnionRect, ValidateRect,
+    ValidateRgn, WindowFromDC
   - Add BITMAP, BITMAPINFOHEADER, BITMAPFILEHEADER structs
 
 ## 2.0.0-nullsafety.7